==> 2D_example_GlycoscienceDB.ipynb <==
import numpy as np
import pandas as pd
import os
from tqdm import tqdm
import torch
import torch.nn as nn

import torch.nn.functional as F
import dgl
import dgl.nn as dglnn
import torch.nn as nn
from dgl import AddSelfLoop

from model_2d.NMR_gcn import NMR_GCN

from train_evaluate_2d import NMR_prediction

from preprocess.create_node_embedding.node_embeddings_glycoscience import create_node_embeddings

from preprocess.build_adjaency_matrix.create_adjaency_matrix_glycoscience import build_adjacency_matrix

from preprocess.create_graph.create_graph_data_glycosciencedb import create_graph

# from dgl import save_graphs, load_graphs
# from tqdm import tqdm
from dgl import save_graphs, load_graphs


##### 1, Create Node embeddings

##### Initialize atom-level node embeddings based on extracted atom and monosaccharide features. 

##### Due to the inconsistent of data formating, as GlycoscienceDB is retrieved from various labs, the method of creating node embedding is slighly different from GODESS.

1, atom_dim: dimension of the atom name that indicate ring position, Eg: 04, C4, C3...

2, mono_dim: dimension of the monosaccharide name, glc.

3, ab_dim: dimension of the anomeric orientation of hydroxyl group.

4, dl_dim: dimension of the fischer convention, fischer d, l.

5, pf_dim: dimension of the number of carbon p, f.

6, residual_dim: dimension of the residual number starting from the root as number 1.



atom_dim = 256

residual_dim = 128

mono_dim = 64
ab_dim = 64
dl_dim = 64
pf_dim = 64

seed=9721

C = create_node_embeddings(seed=seed)

df_atom_embedding, df_residual_embedding, df_monosaccharide_embedding, df_ab_embedding, df_dl_embedding, df_pf_embedding =\
C.create_all_embeddings(atom_dim=atom_dim, residual_dim=residual_dim, mono_dim=mono_dim, 
                       ab_dim=ab_dim, dl_dim=dl_dim, pf_dim=pf_dim)



df_atom_embedding.to_csv(C.out_atom_embed, index=False)
df_residual_embedding.to_csv(C.out_residual_embed, index=False)
df_monosaccharide_embedding.to_csv(C.out_monosaccharide_embed, index=False)

df_ab_embedding.to_csv(C.out_bound_ab, index = False) 
df_dl_embedding.to_csv(C.out_bound_dl, index = False) 
df_pf_embedding.to_csv(C.out_carbon_pf, index = False) 

##### 2, Create adjaency matrix

##### Build adjaency matrix based on threshold distance between atoms

1, threshold_carbon: we construct edges between two carbon atom if their distance is within 1.65Å. 

2, threshold_hydrogen: we construct edges between two hydrogen atom if their distance is within 1.18Å. 

3, threshold_general: for all other atoms, we construct edges between distance two atom is within 1.5Å. 

4, threshold_interaction: we construct interaction edges between atoms if their distance is with in 5Å. 


B = build_adjacency_matrix()

B.calculate_all_matrix()

##### 3, Create glycan graph

##### create glycan graph from prebuilt node embeddings and adjaency matrix, every glycan is considered as a graph and every atom is consided as a node

1, num_test: number of testing glycans, we randomly select 20% of the glycans as test set


num_test = 60 # 299 * 20% 

Create = create_graph(num_test=num_test, seed=seed, use_mono=0, use_ab=0, use_dl=0, use_pf=0)
g, test_index = Create.create_all_graph()

g

##### 4, Train a 4-layer GNN model

num_epoch = 1000
lr = 1e-2


# in_size = atom_dim + mono_dim + ab_dim + dl_dim + pf_dim
in_size = atom_dim

hidden_size_1 = int(in_size / 2)
hidden_size_2 = 256
hidden_size_3 = 128
hidden_size_4 = 32

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

g = g.to(device)
features = g.ndata["feat"]
labels = g.ndata["shift_value"]
# masks = g.ndata['train_mask'], g.ndata['test_mask']


# masks = g.ndata['train_hydrogen_mask'], g.ndata['test_hydrogen_mask']
masks = g.ndata['train_carbon_mask'], g.ndata['test_carbon_mask']
# masks = g.ndata['train_mask'], g.ndata['test_mask']

print(features.dtype)
print(labels.dtype)
# model = NMR_GCN(in_size=576, hid_size=[256, 128, 64, 32], out_size=1).to(device)
model = NMR_GCN(in_size=in_size, hid_size=[hidden_size_1, hidden_size_2,
                                       hidden_size_3, hidden_size_4], out_size=1).to(device)
# model training

# NMR_prediction = NMR_prediction(results_dir='experimental_data_combined/graph_combined/experimental_results_all_hydrogen.csv',
#                                 model_dir='experimental_data_combined/graph_combined/Model_hydrogen_experiment_training.pt',
#                                num_epoch = num_epoch, 
#                                lr = lr)


NMR_prediction = NMR_prediction(results_dir='glycosciencedb/results/training_carbon.csv',
                                results_dir_test = 'glycosciencedb/results/testing_carbon.csv',
                                model_dir='glycosciencedb/results/Model_Godess_carbon.pt', 
                               num_epoch = num_epoch, 
                               lr = lr)


print("Training...")
NMR_prediction.train(g, features, labels, masks, model)

# test the model
print("Testing...")
saved_model = NMR_GCN(in_size=in_size, hid_size=[hidden_size_1, hidden_size_2,
                                                 hidden_size_3, hidden_size_4], out_size=1).to(device)
saved_model.load_state_dict(torch.load(NMR_prediction.model_dir))

# acc = NMR_prediction.evaluate(g, features, labels, masks[0], saved_model, print_out=True)
acc1 = NMR_prediction.evaluate(g, features, labels, masks[0], saved_model, save_train=True, save_test = False)

acc2 = NMR_prediction.evaluate(g, features, labels, masks[1], saved_model, save_train=False, save_test = True)


print("train RMSE {:.4f}".format(acc1))
print("test RMSE {:.4f}".format(acc2))


# acc1 = NMR_prediction.evaluate(g, features, labels, masksc[0], saved_model, save_train=True, save_test = False)

# acc2 = NMR_prediction.evaluate(g, features, labels, masksc[1], saved_model, save_train=False, save_test = True)


# print("train RMSE {:.4f}".format(acc1))
# print("test RMSE {:.4f}".format(acc2))
