==> Step_2.2_tranfer_learning_preprocess_initialize_node_embeddings_godess.ipynb <==
import numpy as np
import pandas as pd
import os
from tqdm import tqdm
import torch
import torch.nn as nn

# from train_evaluate import NMR_prediction

import torch.nn.functional as F
import dgl
import dgl.nn as dglnn
import torch.nn as nn
from dgl import AddSelfLoop

from model_2d.NMR_gcn import NMR_GCN

from train_evaluate_2d import NMR_prediction

from preprocess.create_node_embedding.node_embeddings_godess import create_node_embeddings

from preprocess.build_adjaency_matrix.create_adjaency_matrix_godess import build_adjacency_matrix

from preprocess.create_graph.create_graph_data_godess import create_graph

# from dgl import save_graphs, load_graphs
# from tqdm import tqdm
from dgl import save_graphs, load_graphs


# define the lengtht of the embeddings of each feature
atom_name_dim=256

bound_orig_dim=32

atom_type_dim=32

ab_dim=64

dl_dim=64

pf_dim=64

mono_accurate_dim=128

mono_simple_dim=128

me_dim=32

ser_dim=32

s_dim=64

ac_dim=64

gc_dim=32

seed=9721

C = create_node_embeddings(data_dir = 'godess/data_reformulate/',
                           out_atom_name_embed='godess/node_embedding_reformulate/atom_name_embed.csv',
                           out_bound_orig_embed='godess/node_embedding_reformulate/bound_orig.csv',
                           out_atom_type_embed='godess/node_embedding_reformulate/atom_type.csv',
                           out_bound_AB_embed='godess/node_embedding_reformulate/bound_ab.csv',
                           out_DL_embed='godess/node_embedding_reformulate/bound_dl.csv',
                           out_PF_embed='godess/node_embedding_reformulate/carbon_pf.csv',
                           out_monosaccharide_accurate_embed='godess/node_embedding_reformulate/monosaccharide_accurate_embed.csv',
                           out_monosaccharide_simple_embed='godess/node_embedding_reformulate/monosaccharide_simple_embed.csv',
                           out_me_embed='godess/node_embedding_reformulate/root_me_embed.csv',
                           out_ser_embed='godess/node_embedding_reformulate/root_ser_embed.csv',
                           out_s_embed='godess/node_embedding_reformulate/component_s_embed.csv',
                           out_ac_embed='godess/node_embedding_reformulate/component_ac_embed.csv',
                           out_gc_embed='godess/node_embedding_reformulate/component_gc_embed.csv',
                           seed=seed)
C.write_all_embeddings(atom_name_dim = atom_name_dim, bound_orig_dim = bound_orig_dim, 
                           atom_type_dim = atom_type_dim, ab_dim = ab_dim, dl_dim = dl_dim,
                           pf_dim = pf_dim, mono_accurate_dim = mono_accurate_dim, 
                           mono_simple_dim = mono_simple_dim, me_dim=me_dim, 
                          ser_dim=ser_dim, s_dim=s_dim, ac_dim=ac_dim, gc_dim=gc_dim)
