==> Step_2.3_tranfer_learning_preprocess_merge_two_node_embeddings.ipynb <==
import numpy as np
import pandas as pd
import os
from tqdm import tqdm
import torch
import torch.nn as nn

# glycosciences
glycoscience_atom_embed = pd.read_csv('glycosciencedb/node_embedding_reformulate/atom_embed.csv')

glycoscience_monosaccharide_embed = pd.read_csv('glycosciencedb/node_embedding_reformulate/monosaccharide_embed.csv')

glycoscience_bound_AB_embed = pd.read_csv('glycosciencedb/node_embedding_reformulate/ab_embed.csv')

glycoscience_DL_embed = pd.read_csv('glycosciencedb/node_embedding_reformulate/dl_embed.csv')

glycoscience_PF_embed = pd.read_csv('glycosciencedb/node_embedding_reformulate/pf_embed.csv')

# godess

godess_atom_embed=pd.read_csv('godess/node_embedding_reformulate/atom_name_embed.csv')

godess_monosaccharide_embed=pd.read_csv('godess/node_embedding_reformulate/monosaccharide_simple_embed.csv')

godess_bound_AB_embed=pd.read_csv('godess/node_embedding_reformulate/bound_ab.csv')

godess_DL_embed=pd.read_csv('godess/node_embedding_reformulate/bound_dl.csv')

godess_PF_embed=pd.read_csv('godess/node_embedding_reformulate/carbon_pf.csv')


merged_atom_embed_path='merged_embed/atom_name_embed.csv'

merged_monosaccharide_embed_path='merged_embed/monosaccharide_embed.csv'

merged_bound_AB_embed_path='merged_embed/ab_embed.csv'

merged_DL_embed_path='merged_embed/dl_embed.csv'

merged_PF_embed_path='merged_embed/pf_embed.csv'

def merge_two_df(df1, df2):

    common_columns = df1.columns.intersection(df2.columns)
    avg_df = (df1[common_columns] + df2[common_columns]) / 2

    df1_unique_column = np.array([item for item in df1.columns if item not in common_columns])
    df1_unique = df1[df1_unique_column]

    df2_unique_column = np.array([item for item in df2.columns if item not in common_columns])
    df2_unique = df2[df2_unique_column]

    df_new = pd.concat([avg_df, df1_unique, df2_unique], axis = 1)
    
    return df_new

df_merged_atom_embed = merge_two_df(glycoscience_atom_embed, godess_atom_embed)
df_merged_atom_embed.to_csv(merged_atom_embed_path)

df_merged_monosaccharide_embed = merge_two_df(glycoscience_monosaccharide_embed, godess_monosaccharide_embed)
df_merged_monosaccharide_embed.to_csv(merged_monosaccharide_embed_path)

df_merged_bound_AB_embed = merge_two_df(glycoscience_bound_AB_embed, godess_bound_AB_embed)
df_merged_bound_AB_embed.to_csv(merged_bound_AB_embed_path)

df_merged_DL_embed = merge_two_df(glycoscience_DL_embed, godess_DL_embed)
df_merged_DL_embed.to_csv(merged_DL_embed_path)

df_merged_PF_embed = merge_two_df(glycoscience_PF_embed, godess_PF_embed)
df_merged_PF_embed.to_csv(merged_PF_embed_path)

df_merged_atom_embed.shape, df_merged_monosaccharide_embed.shape, df_merged_bound_AB_embed.shape, df_merged_DL_embed.shape, df_merged_PF_embed.shape

df_merged_atom_embed.columns
